+++ spynnaker/pyNN/models/utility_models/delays/delay_extension_vertex.py
@@ -25,8 +25,7 @@ from pacman.model.constraints.partitioner_constraints import (
 from pacman.model.resources import (
     ConstantSDRAM, CPUCyclesPerTickResource, DTCMResource, ResourceContainer)
 from spinn_front_end_common.abstract_models import (
-    AbstractGeneratesDataSpecification,
-    AbstractProvidesOutgoingPartitionConstraints)
+    AbstractGeneratesDataSpecification)
 from spinn_front_end_common.abstract_models.impl import (
     TDMAAwareApplicationVertex)
 from spinn_front_end_common.interface.simulation import simulation_utilities
@@ -60,7 +59,7 @@ _MAX_OFFSET_DENOMINATOR = 10
 
 class DelayExtensionVertex(
         TDMAAwareApplicationVertex, AbstractGeneratesDataSpecification,
-        AbstractProvidesOutgoingPartitionConstraints, AbstractHasDelayStages):
+        AbstractHasDelayStages):
     """ Provide delays to incoming spikes in multiples of the maximum delays\
         of a neuron (typically 16 or 32)
     """
@@ -429,11 +428,6 @@ class DelayExtensionVertex(
         """
         words_per_atom = 11 + 16
         return words_per_atom * BYTES_PER_WORD * vertex_slice.n_atoms
-
-    @overrides(AbstractProvidesOutgoingPartitionConstraints.
-               get_outgoing_partition_constraints)
-    def get_outgoing_partition_constraints(self, partition):
-        return [ContiguousKeyRangeContraint()]
 
     def delay_generator_data(self, vertex_slice):
         if vertex_slice in self.__delay_generator_data:
